Added process info to output. Edited functions' names.
