+++ gps_anomaly/config_heading.py
@@ -1,5 +1,8 @@
-class Angle:
-    header_limit = 10
+class Heading:
+    """
+        header_limit : define the distance limit for detecting anomaly
+    """
+    header_limit = 30
 
     def __init__(self):
         self.upper_angle = int
